+++ src/utils/components/hover.ts
@@ -2,6 +2,7 @@ import TelemetryReporter from "@vscode/extension-telemetry";
 import * as vscode from "vscode";
 import {
   WATERMELON_HISTORY_COMMAND,
+  WATERMELON_OPEN_LINK_COMMAND,
   WATERMELON_PULLS_COMMAND,
 } from "../../constants";
 import getNumberOfFileChanges from "../getNumberOfFileChanges";
@@ -34,15 +35,26 @@ const hover = ({ reporter }: { reporter: TelemetryReporter | null }) => {
           JSON.stringify(args)
         )}`
       );
+      const mailtoLinkCommandUri = vscode.Uri.parse(
+        `command:${WATERMELON_OPEN_LINK_COMMAND}?${encodeURIComponent(
+          JSON.stringify({
+            url: `mailto:${latestCommit.authorEmail}`,
+            source: "hover",
+          })
+        )}`
+      );
       const content = new vscode.MarkdownString(
         `$(git-pull-request)[Understand the code context](${startCommandUri}) with Watermelon 🍉`
       );
       content.appendMarkdown(`\n\n`);
       content.appendMarkdown(
-        `The latest commit is "${latestCommit.message}" by **${
+        `The latest commit was made by [$(mail)${
           latestCommit.authorName
-        }** on **${latestCommit.commitDate.toLocaleDateString()}**`
+        }](${mailtoLinkCommandUri}) on **${latestCommit.commitDate.toLocaleDateString()}**:
+        `
       );
+      content.appendMarkdown(`\n`);
+      content.appendMarkdown(latestCommit.message);
       content.appendMarkdown(`\n\n`);
       content.appendMarkdown(
         `$(git-commit)[View the history for this line](${blameCommandUri}) with Watermelon 🍉`