+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to the "watermelon" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.5.0]
+
+- Commit messages now allow markdown, including lists
+- Added a Draft PR indicator 
+- Sidebar UI reorganized to show a single commit and a single PR in a single view
+
+## [1.4.6]
+
+- New `watermelon.recommend` command, to recommend a watermelon to everyone on the repo
+- New `watermelon.login` command, to login using GitHub from your keyboard
+- We now track every time the `watermelon.show` command is called
+- You can now email the committer from the hover
+- Changed how the buttons on the webview work to match commands
+
 ## [1.4.5]
 
 - Better daily summary UI